Add tests for helper methods
I'd like to encourage testing from Day 1 onsite, rather than Day 2,
as is done in Chicago.  I'm providing some tests for the small
helper methods that students are directed to write in Release 0.

Unfortunately, the test setup involves stubbing gets and puts calls
which hit the main object, so the test setup is probably a bit much
for Day-1 students, but I've tried to provide thorough comments on
what is happening in the tests.

There are no tests for the deaf_grandma method.  Students are not
instructed to write any.  They would be more complex than I
believe they could write.